feat(guides): push to repo
